Fix strided tensor views for fused projections
